+++ src/kernel/allocator/LinkedListAllocator.h
@@ -35,7 +35,7 @@ public:
 
       void init ();        
       void dump_free_memory ();
-      void* alloc (unsigned int req_size);
+      void* alloc (uint64_t req_size);
       void free (void *ptr);
 
 };
